+++ app/src/main/kotlin/ac/mdiq/podcini/config/ClientConfig.kt
@@ -31,8 +31,6 @@ import kotlinx.coroutines.launch
 
 
 object ClientConfig {
-//    var USER_AGENT: String? = null
-
     private var initialized = false
     var nmJob: Job? = null